requirements/base.txt
celery==4.2.0
django-filter==2.0.0
django-money==0.14.3
Django==2.1.3
djangorestframework==3.9.0
feedparser==5.2.1
psycopg2==2.7.6.1
